requirements.txt
openai
ffmpeg-python
openai-whisper
streamlit
pytube
youtube_transcript_api